End to end tests (with support infrastructure) YAY!
Put together a pair of end to end tests (one assuming cwd has state
files, the other using -dir) of the --list functionality.  Still lots to
be done, but....

Added an inventory package that defines a type Inventory and has a
function that will create a new one from the long-form dynamic inventory
JSON (the form we generate, at least...).